Update to account for endCursor having value
Previously, if the results were on the last page, the endCursor value would be null, which technically isn’t correct. The endCursor should, in fact, display the cursor of the last node in the page. The `hasNextPage` value, however, will still be false.

The endCursor will still be null if no results could be found.